fix: Fixed resize function to avoid creating auxiliar dynamic arrays that produce errors
Fixed resize function which created memory errors when assigning information to the new array, which is never returned properly to the fucntion from it was called. For this reason, the function "realloc" was used instead as it doesn't need to create an auxiliar array to be used and has the same functionality.